feat: API Context implemented and tests erased
