+++ src/components/MainPage/MainPage.tsx
@@ -1,4 +1,3 @@
-import TodoProvider, { TodoContext, TodoContextType } from "../../context/main";
 import * as React from 'react';
 import { Banner } from "../Banner";
 import { MOCK_ALLPRODUCTS } from "../Header/Mock";
@@ -11,10 +10,8 @@ interface MainPageProps {
 }
 
 export function MainPage ({AddToCart}: MainPageProps) {
-    const { saveTodo } = React.useContext(TodoContext) as TodoContextType;
-    console.log(saveTodo);
     return (
-        <MainPageContainer>
+        <MainPageContainer>        
             <Banner />
             <h2 className="ProductGroupTitle">Nossos cafés</h2>
             <div className="ProductGroup">
