+++ src/pages/checkout.tsx
@@ -3,34 +3,12 @@ import React, { useState } from 'react'
 import { Cart } from '../components/Cart'
 import { FormCard } from '../components/FormCard'
 import { Header } from '../components/Header'
-import { MOCK_ALLPRODUCTS, MOCK_CART} from '../components/Header/Mock'
 import { PaymentCard } from '../components/PaymentCard'
-import { CartProduct } from '../models/Cart'
+import { CartContext, CartContextType } from '../context/CartContext'
 import { BodyContainer } from '../styles/pages/homeStyles'
 
 const Home: NextPage = () => {
-  const [currentCart, setCurrentCart] = useState(MOCK_CART as CartProduct[]);
-  function AddToCart (productId: number, addedQuantity: number, newQuantity = -1){
-    let ProductIsNotInCart:boolean;    
-    ProductIsNotInCart=true; 
-    
-    const newQuantitiesCart = currentCart.map(
-      (productCart)=>{
-        if (productId==productCart.product.id){
-          if (newQuantity<0) {productCart.quantity+=addedQuantity;}
-          else {productCart.quantity=newQuantity;}
-          ProductIsNotInCart=false;
-        }
-        return productCart;
-      });
-    const CartWithoutZeroedProducts =  newQuantitiesCart.filter((product) => {return (product.quantity>0)});
-    
-    if (ProductIsNotInCart) {      
-      CartWithoutZeroedProducts.push({product: MOCK_ALLPRODUCTS[productId], quantity: addedQuantity});
-    }
-    setCurrentCart(CartWithoutZeroedProducts);
-  }
-  
+  const { currentCart,AddToCart } = React.useContext(CartContext) as CartContextType;  
   const [paymentType, SetPaymentType] = useState("credit" as string); 
 
   return (
